# openapi.yaml
openapi: 3.0.0
info:
  title: Indikara Backend API
  version: 1.0.0
  description: API for authentication, products, and health checks
servers:
  - url: https://backend-wei5.onrender.com
    description: Deployment server
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        isAdmin:
          type: boolean
        isDemo:
          type: boolean
...
    OrderProductItem:
      type: object
      properties:
        product:
          type: string
          description: Product ObjectId or SKU
        quantity:
          type: number
        price:
          type: number
    Order:
      type: object
      properties:
        _id:
          type: string
        txnid:
          type: string
        user:
          $ref: '#/components/schemas/User'
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProductItem'
        shippingAddress:
          type: object
        totalPrice:
          type: number
        paymentMethod:
          type: string
        paymentResult:
          type: object
          description: Payment gateway response / reference
        isPaid:
          type: boolean
        paidAt:
          type: string
          format: date-time
        status:
          type: string
...
    CreateOrderProductItem:
      type: object
      properties:
        product:
          type: string
          description: Product ObjectId or SKU
        quantity:
          type: number
    CreatePendingRequest:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderProductItem'
        shippingAddress:
          type: object
      required:
        - products
    CreatePendingResponse:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
        txnid:
          type: string
...
paths:
...
  /api/orders:
    post:
      summary: Create a new order
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
  /api/orders/create-pending:
    post:
      summary: Create a pending order and get a txnid (used for PayU)
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePendingRequest'
      responses:
        '201':
          description: Pending order created with txnid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePendingResponse'
              examples:
                demo:
                  value:
                    order:
                      txnid: tx_12345
                      totalPrice: 100
                    txnid: tx_12345
  /api/orders/payu-success:
    post:
      summary: Create an order after PayU redirects/callbacks (post-payment)
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created
  /api/orders/{id}:
    get:
      summary: Get order by ID
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /api/orders/my:
    get:
      summary: Get orders for the logged-in user
      tags:
        - Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...